+++ Integration/Romberg.m
@@ -28,7 +28,7 @@ arguments
     f (1,1) function_handle
     a (1,1) double
     b (1,1) double
-    opts.tol (1,1) int = 1e-10
+    opts.tol (1,1) double = 1e-10
     opts.maxiter (1,1) int = 100
 end
